Update the common layout tests
