+++ packages/core/admin/admin/src/pages/MarketplacePage/tests/plugins.test.js
@@ -56,7 +56,7 @@ const waitForReload = async () => {
   );
 };
 
-describe('Plugins tab', () => {
+describe('Marketplace page - plugins tab', () => {
   let renderedContainer;
   let history;
 
@@ -380,5 +380,25 @@ describe('Plugins tab', () => {
 
     userEvent.click(screen.getByRole('tab', { name: /plugins/i }));
     expect(collectionCards.length).toBe(2);
+  });
+
+  it('shows the correct options on sort select', () => {
+    const sortButton = screen.getByRole('button', { name: /Sort by/i });
+    userEvent.click(sortButton);
+
+    const alphabeticalOption = screen.getByRole('option', { name: 'Alphabetical order' });
+    const newestOption = screen.getByRole('option', { name: 'Newest' });
+
+    expect(alphabeticalOption).toBeVisible();
+    expect(newestOption).toBeVisible();
+  });
+
+  it('changes the url on sort option select', () => {
+    const sortButton = screen.getByRole('button', { name: /Sort by/i });
+    userEvent.click(sortButton);
+
+    const newestOption = screen.getByRole('option', { name: 'Newest' });
+    userEvent.click(newestOption);
+    expect(history.location.search).toEqual('?sort=submissionDate:desc');
   });
 });
